CI: remove skip_s3_upload option
This option was broken. Its logic, for when this flag is provided, was
clearly reversed. The only reason we didn't encounter issues was that
the flag is usually empty.

We've never actually needed this flag. It was something I thought would
be nice early on. It's more clearly correct to remove this option.